refactor: extract mod into file
